Newton solver is changed from trust region to line search
Specific changes are as follows.

- file name change:

tron.cpp and tron.h are changed to newton.cpp and newton.h, respectively.

In newton.cpp, the subroutine trpcg is changed to pcg.

Some functions (in the class newton) used for trust region are removed.

- classes of functions in linear.cpp:

A new class of function l2r_erm_fun is added to cover lr and svm.
It assumes the following function form

min_w w^Tw/2 + \sum C_i \xi(w^Tx_i), where \xi() is the loss

Some common utilities in particular the line-search subroutine (see below) are
implemented here.

- line search subroutines:

A special line-search subroutine for l2r_erm_fun is implemented in linear.cpp
so for each step size the function value can be cheaply calculated.

A base implementation of line search is provided in newton.cpp, where it
calculates every function value from scratch.

We require that the line-search subroutine to update w and also maintain
the function value.

- CG stopping criterion

It's changed to check a quadratic approximation instead of the
residual. See the release notes of version 2.40.

- CG stopping tolerance

LIBLINEAR enlarged eps_cg if the initial primal solution is not
null. This was from Chu et al. (2015) to avoid too many CG steps
in warm start for parameter selection. We found that this setting
is no longer needed; see details in version 2.40 release notes.